Outsourced the individual match viewer
