+++ src/components/game_specific/SimulatorMatchViewer/2022.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import SimulatorMatchViewer_Universal from "./_Universal";
+
+export default function SimulatorMatchViewer({ sim }) {
+    return <SimulatorMatchViewer_Universal sim={sim}/>
+}
